backend/requirements.txt
Django==3.2.25
djongo>=1.3.6
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
python-dotenv>=1.0.1
pymongo<4.0.0
whitenoise>=6.6.0
gunicorn>=21.2.0
sqlparse==0.2.4
numpy
pandas
dnspython
tensorflow
pandas_ta
certifi
yfinance==0.2.40
# Backend / Async / Scheduling
celery
redis
django-celery-results
django-celery-beat

# MT5 & Trading
MetaTrader5

# Scientific & Optimization
scipy
statsmodels
scikit-learn

# Time & Date Handling
pytz
python-dateutil

# Serialization & Security
cryptography

# WebSockets
channels
channels-redis

